backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
chromadb
pillow
pymupdf
python-multipart
pydantic
numpy
paddlepaddle
paddleocr
